# server/docs/swagger.yaml
basePath: /
definitions:
  database.DownloadPreference:
    properties:
      info_hash_preference:
        $ref: '#/definitions/database.DownloadPreferenceInfoHash'
      match_type:
        type: string
      string_match_preference:
        $ref: '#/definitions/database.DownloadPreferenceString'
    type: object
  database.DownloadPreferenceInfoHash:
    properties:
      info_hash:
        type: string
    type: object
  database.DownloadPreferenceString:
    properties:
      case_sensitive:
        type: boolean
      match_string:
        type: string
    type: object
...
  database.MediaFile:
    properties:
      bit_rate:
        type: string
      created_at:
        type: string
      duration:
        type: integer
      file_id:
        type: integer
      file_idx:
        type: integer
      file_name:
        type: string
      file_origin:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      format_long_name:
        type: string
      format_name:
        type: string
      framerate:
        type: string
      height:
        type: integer
      original_file_name:
        type: string
      record_id:
        type: integer
      source_uri:
        type: string
      streams:
        items:
          $ref: '#/definitions/database.Stream'
        type: array
      updated_at:
        type: string
      width:
        type: integer
    type: object
...
  database.Stream:
    properties:
      channel_layout:
        type: string
      channels:
        type: integer
      codec_long_name:
        type: string
      codec_name:
        type: string
      codec_type:
        description: video, audio, subtitle
        type: string
      color_primaries:
        type: string
      color_range:
        type: string
      color_space:
        type: string
      color_transfer:
        type: string
      language:
        type: string
      pix_fmt:
        type: string
      profile:
        type: string
      title:
        type: string
    type: object
...
  database.VideoMetadata:
    properties:
      bit_rate:
        type: string
      duration:
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      format_long_name:
        type: string
      format_name:
        type: string
      framerate:
        type: string
      height:
        type: integer
      streams:
        items:
          $ref: '#/definitions/database.Stream'
        type: array
      width:
        type: integer
    type: object
...
  providers.ProviderObject:
    properties:
      provider:
        description: provider name in /providers folder
        type: string
      streams:
        items:
          $ref: '#/definitions/providers.StreamObject'
        type: array
    type: object
  providers.ProviderResponseObject:
    properties:
      episode_number:
        type: integer
      episode_source_id:
        description: tv shows only
        type: string
      imdb_id:
        description: starts with 'tt'
        type: string
      media_source:
        type: string
      media_type:
        description: movies or tvshows, etc.
        type: string
      providers:
        items:
          $ref: '#/definitions/providers.ProviderObject'
        type: array
      season_number:
        description: shows only
        type: integer
      source_id:
        type: string
    type: object
  providers.StreamObject:
    properties:
      description:
        type: string
      encoded_data:
        description: data encoded in AES for playing streams in hound
        type: string
      file_idx:
        description: file index for p2p type
        type: integer
      file_name:
        description: might not be reliable
        type: string
      file_size:
        description: file size in bytes
        type: integer
      info_hash:
        type: string
      provider:
        type: string
      sources:
        description: trackers for p2p
        items:
          type: string
        type: array
      stream_protocol:
        description: http or p2p
        type: string
      title:
        type: string
      uri:
        description: magnet link, http link, or file path
        type: string
      video_metadata:
        $ref: '#/definitions/database.VideoMetadata'
    type: object
...
  v1.CancelIngestTaskResponse:
    properties:
      ingest_task_id:
        type: integer
      status:
        example: pending_cancel
        type: string
    type: object
...
  v1.DownloadResponse:
    properties:
      status:
        example: started
        type: string
    type: object
...
  v1.TVSeasonDownloadRequest:
    properties:
      episodes_to_download:
        items:
          type: integer
        type: array
      preference_list:
        items:
          $ref: '#/definitions/database.DownloadPreference'
        type: array
      skip_downloaded_episodes:
        type: boolean
      strict_match:
        description: whether if no match is found, to fail the download
        type: boolean
    type: object
  v1.TVSeasonDownloadResponse:
    properties:
      queued_episodes:
        items:
          type: integer
        type: array
      season_number:
        type: integer
      skipped_episodes:
        items:
          $ref: '#/definitions/v1.skippedEpisode'
        type: array
      status:
        example: queued
        type: string
    type: object
  v1.V1ErrorResponse:
    properties:
      error:
        example: error message
        type: string
      status:
        enum:
        - success
        - error
        example: error
        type: string
    type: object
  v1.V1SuccessResponse:
    properties:
      data: {}
      status:
        enum:
        - success
        - error
        example: success
        type: string
    type: object
  v1.skippedEpisode:
    properties:
      episode_number:
        type: integer
      error:
        type: string
    type: object
...
  view.MediaFilesResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/database.MediaFile'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total_records:
        type: integer
    type: object
...
host: localhost:2323
info:
  contact: {}
  description: |-
    V1 Hound Media Server API Documentation
    While in beta, please expect breaking api changes in short/no notice
  title: Hound API Documentation V1
  version: beta-0.1.0
paths:
...
  /api/v1/download/{encodedString}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Encoded Stream String - get from Query providers
        in: path
        name: encodedString
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DownloadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Download media file to server
      tags:
      - Download
...
  /api/v1/ingest/{taskID}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CancelIngestTaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Cancel an ingest/download task
      tags:
      - Download
  /api/v1/media_files:
    get:
      consumes:
      - application/json
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/view.MediaFilesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get all media file records
      tags:
      - Media Files
  /api/v1/media_files/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Media File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Delete a media file
      tags:
      - Media Files
...
  /api/v1/movie/{id}/media_files:
    get:
      consumes:
      - application/json
      description: Get
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/providers.ProviderResponseObject'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get Movie Media Files by ID
      tags:
      - Media Files
...
  /api/v1/tv/{id}/media_files:
    get:
      consumes:
      - application/json
      parameters:
      - description: TV Show ID
        in: path
        name: id
        required: true
        type: integer
      - description: Season Number
        in: query
        name: season
        type: integer
      - description: Episode Number
        in: query
        name: episode
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/providers.ProviderResponseObject'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Search TV Show Media Files by ID
      tags:
      - Media Files
...
  /api/v1/tv/{id}/season/{seasonNumber}/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Season Number
        in: path
        name: seasonNumber
        required: true
        type: integer
      - description: Download Preferences
        in: body
        name: request
        schema:
          $ref: '#/definitions/v1.TVSeasonDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.TVSeasonDownloadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Download TV Season
      tags:
      - Download
...
swagger: "2.0"